Add down function to revert to old changes

+++ db/migrate/20220715081950_remove_profile_pic_from_user.rb
@@ -1,5 +1,15 @@
 class RemoveProfilePicFromUser < ActiveRecord::Migration[7.0]
-  def change
-    remove_column(:users, :profile_pic, :text)
+  def change 
+    def up
+      if column_exists?(:doctors, :certificate)
+        remove_column(:users, :profile_pic, :text)
+       end
+     end
+     
+     def down
+      if !column_exists?(:doctors, :certificate)
+        add_column :users, :profile_pic, :text
+       end
+     end
   end
 end
